lib: move glob from path to internal/fs/glob
