+++ lib/internal/test_runner/coverage.js
@@ -25,7 +25,7 @@ const {
   readFileSync,
   rmSync,
 } = require('fs');
-const { setupCoverageHooks, isWindows, isMacOS } = require('internal/util');
+const { setupCoverageHooks } = require('internal/util');
 const { tmpdir } = require('os');
 const { join, resolve, relative } = require('path');
 const { fileURLToPath } = require('internal/url');
@@ -36,30 +36,13 @@ const {
     ERR_SOURCE_MAP_MISSING_SOURCE,
   },
 } = require('internal/errors');
+const { glob } = require('internal/fs/glob');
+
 const kCoverageFileRegex = /^coverage-(\d+)-(\d{13})-(\d+)\.json$/;
 const kIgnoreRegex = /\/\* node:coverage ignore next (?<count>\d+ )?\*\//;
 const kLineEndingRegex = /\r?\n$/u;
 const kLineSplitRegex = /(?<=\r?\n)/u;
 const kStatusRegex = /\/\* node:coverage (?<status>enable|disable) \*\//;
-
-let minimatch;
-function lazyMinimatch() {
-  minimatch ??= require('internal/deps/minimatch/index');
-  return minimatch;
-}
-
-function glob(path, pattern, windows) {
-  return lazyMinimatch().minimatch(path, pattern, {
-    __proto__: null,
-    nocase: isMacOS || isWindows,
-    windowsPathsNoEscape: true,
-    nonegate: true,
-    nocomment: true,
-    optimizationLevel: 2,
-    platform: windows ? 'win32' : 'posix',
-    nocaseMagicOnly: true,
-  });
-}
 
 class CoverageLine {
   constructor(line, startOffset, src, length = src?.length) {
